+++ src/io/configFile.cpp
@@ -1,6 +1,6 @@
 /*
 author          Oliver Blaser
-date            27.01.2022
+date            28.01.2022
 copyright       MIT - Copyright (c) 2022 Oliver Blaser
 */
 
@@ -19,9 +19,33 @@ namespace
 
 
 /*!
+* \class omw::io::ConfigFile
+* 
+* `#include <omw/io/configFile.h>`
+*
+* \section class_omw__io_ConfigFile_section_fileSpec File Specification
+* 
+* ...
+*/
+
+omw::io::ConfigFile::ConfigFile()
+{}
+
+omw::io::ConfigFile::~ConfigFile()
+{}
+
+
+
+/*!
 * \class omw::io::IniFileM2M
 * 
 * `#include <omw/io/configFile.h>`
 *
 * Does not store (in RAM) unnecessary information, therefore the saved file is kind of minimized.
 */
+
+omw::io::IniFileM2M::IniFileM2M()
+{}
+
+omw::io::IniFileM2M::~IniFileM2M()
+{}