+++ release-notes.md
@@ -14,6 +14,7 @@ Refactor
 New
  - `omw::preview::crc16_kermit()` (checksum.h)
  - File interface classes (file.h)
+ - `omw::io::ConfigFile` and `omw::io::IniFileM2M` (configFile.h)